requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
scikit-learn
joblib
pandas
python-multipart
requests